feat: notFound and fallback error pages
