+++ src/core/translation/locales/es/general.json
@@ -23,6 +23,15 @@
     "expiry": "Vencimiento",
     "ending": "termina en"
   },
-  "router-not-found": "Página no encontrada :(",
-  "general-success-continue": "Continuar"
+  "general-success-continue": "Continuar",
+  "general-error-not-found": {
+    "header": "404 - Página no encontrada",
+    "subheader": "¡Vaya! La página que buscas no existe en",
+    "go-to-home-btn": "Ir a Inicio"
+  },
+  "geneal-error-internal": {
+    "header": "Error Interno del Servidor",
+    "subheader": "Nos disculpamos por las molestias. Por favor, inténtelo de nuevo más tarde",
+    "home-btn": "Página Principal"
+  }
 }
