+++ src/core/translation/locales/kr/general.json
@@ -23,6 +23,15 @@
     "expiry": "유효기간",
     "ending": "끝자리"
   },
-  "router-not-found": "페이지를 찾을 수 없습니다 :(",
-  "general-success-continue": "계속"
+  "general-success-continue": "계속",
+  "general-error-not-found": {
+    "header": "404 - 페이지를 찾을 수 없습니다",
+    "subheader": "죄송합니다. 요청하신 페이지를 찾을 수 없습니다",
+    "go-to-home-btn": "홈으로 이동"
+  },
+  "general-error-internal": {
+    "header": "내부 서버 오류",
+    "subheader": "불편을 끼쳐드려 죄송합니다. 나중에 다시 시도해 주세요",
+    "home-btn": "홈 페이지"
+  }
 }
