+++ src/pages/error/fallback/fallback.module.scss
@@ -1,0 +1,34 @@
+@import 'src/styles/constants/_primitives.scss';
+
+.container {
+    position: fixed;
+    width: 100vw;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 1.5rem;
+    padding: 3rem 1.5rem;
+    text-align: center;
+
+    .error {
+        &__code {
+            font-size: 80px;
+            font-weight: 800;
+            color: $color-error-600;
+        }
+
+        &__msg {
+            font-size: 48px;
+            font-weight: 600;
+            color: $color-grey-900;
+        }
+
+        &__details {
+            font-size: 24px;
+            color: $color-grey-700;
+            margin-bottom: 1rem;
+        }
+    }
+}
