--- src/pages/fallback/fallback.module.scss
@@ -1,39 +1,0 @@
-.container {
-    position: fixed;
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-.errorCode {
-    font-size: 5rem;
-    font-weight: 800;
-    color: var(--color-error-01);
-}
-.errorMsg {
-    text-align: center;
-    font-size: var(--font-size-xl);
-    color: var(--color-black-01);
-}
-.errorDetails {
-    text-align: center;
-    font-size: var(--font-size-l);
-    color: var(--color-black-01);
-    margin-top: 1rem;
-}
-.buttonContainer {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 1rem;
-}
-.button {
-    width: 100%;
-    margin: auto;
-}
-.link {
-    color: var(--color-white);
-}
